refactor: merge Twitch/SharedMessageBuilder into MessageBuilder
