+++ src/controllers/commands/builtin/twitch/RemoveModerator.cpp
@@ -6,7 +6,6 @@
 #include "providers/twitch/api/Helix.hpp"
 #include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
-#include "providers/twitch/TwitchMessageBuilder.hpp"
 #include "util/Twitch.hpp"
 
 namespace chatterino::commands {
