+++ src/util/Helpers.hpp
@@ -6,6 +6,7 @@
 
 #include <cmath>
 #include <optional>
+#include <utility>
 #include <vector>
 
 namespace chatterino {
